+++ src/main/java/forestry/api/apiculture/ForestryBeeEffects.java
@@ -32,5 +32,6 @@ public class ForestryBeeEffects {
 	public static final ResourceLocation HAKUNA_MATATA = ForestryConstants.forestry("bee_effect_hakuna_matata");
 	public static final ResourceLocation GLOW_BERRY_GROW = ForestryConstants.forestry("bee_effect_glow_berry_grow");
 	public static final ResourceLocation GUARDIAN = ForestryConstants.forestry("bee_effect_guardian");
+	public static final ResourceLocation PHASING = ForestryConstants.forestry("bee_effect_phasing");
 	public static final ResourceLocation ASCENSION = ForestryConstants.forestry("bee_effect_ascension");
 }