+++ src/main/java/forestry/plugin/DefaultForestryPlugin.java
@@ -69,6 +69,7 @@ import forestry.apiculture.genetics.effects.GuardianBeeEffect;
 import forestry.apiculture.genetics.effects.HeroicBeeEffect;
 import forestry.apiculture.genetics.effects.IgnitionBeeEffect;
 import forestry.apiculture.genetics.effects.MisanthropeBeeEffect;
+import forestry.apiculture.genetics.effects.PhasingBeeEffect;
 import forestry.apiculture.genetics.effects.PotionBeeEffect;
 import forestry.apiculture.genetics.effects.PotionBeeEffectExclusive;
 import forestry.apiculture.genetics.effects.RadioactiveBeeEffect;
@@ -220,6 +221,7 @@ public class DefaultForestryPlugin implements IForestryPlugin {
 		Supplier<List<ItemStack>> frozenComb = getHoneyComb(EnumHoneyComb.FROZEN);
 		Supplier<List<ItemStack>> mossyComb = getHoneyComb(EnumHoneyComb.MOSSY);
 		Supplier<List<ItemStack>> spongeComb = getHoneyComb(EnumHoneyComb.SPONGE);
+		Supplier<List<ItemStack>> simmerComb = getHoneyComb(EnumHoneyComb.SIMMERING);
 
 		apiculture.registerHive(ForestryBeeSpecies.FOREST, HiveDefinition.FOREST)
 				.addDrop(0.80, ForestryBeeSpecies.FOREST, honeyComb, 0.7f)
@@ -261,6 +263,9 @@ public class DefaultForestryPlugin implements IForestryPlugin {
 				.addDrop(0.80, ForestryBeeSpecies.AQUATIC, spongeComb, 0.4F)
 				.addDrop(0.03, ForestryBeeSpecies.VALIANT, spongeComb);
 
+		apiculture.registerHive(ForestryBeeSpecies.EMBITTERED, HiveDefinition.NETHER)
+				.addDrop(0.80, ForestryBeeSpecies.EMBITTERED, simmerComb, 0.7F);
+
 		// Common village bees
 		apiculture.addVillageBee(ForestryBeeSpecies.FOREST, false);
 		apiculture.addVillageBee(ForestryBeeSpecies.MEADOWS, false);
@@ -272,7 +277,10 @@ public class DefaultForestryPlugin implements IForestryPlugin {
 
 		// Rare village bees
 		apiculture.addVillageBee(ForestryBeeSpecies.FOREST, true, Map.of(BeeChromosomes.TOLERATES_RAIN, ForestryAlleles.TRUE));
-		apiculture.addVillageBee(ForestryBeeSpecies.COMMON, true);
+		apiculture.addVillageBee(ForestryBeeSpecies.COMMON, true, Map.of(
+				BeeChromosomes.TEMPERATURE_TOLERANCE, ForestryAlleles.TOLERANCE_BOTH_1,
+				BeeChromosomes.HUMIDITY_TOLERANCE, ForestryAlleles.TOLERANCE_BOTH_1
+		));
 		apiculture.addVillageBee(ForestryBeeSpecies.VALIANT, true);
 
 		// Default flower types
@@ -314,6 +322,7 @@ public class DefaultForestryPlugin implements IForestryPlugin {
 		apiculture.registerBeeEffect(ForestryBeeEffects.HAKUNA_MATATA, new PotionBeeEffectExclusive(false, ApicultureEffects.HAKUNA_MATATA.get(), 20 * 60 * 3, 100, 1.0f, ApicultureEffects.MATATA.get()));
 		apiculture.registerBeeEffect(ForestryBeeEffects.GLOW_BERRY_GROW, new GlowBerryGrowEffect());
 		apiculture.registerBeeEffect(ForestryBeeEffects.GUARDIAN, new GuardianBeeEffect());
+		apiculture.registerBeeEffect(ForestryBeeEffects.PHASING, new PhasingBeeEffect());
 		apiculture.registerBeeEffect(ForestryBeeEffects.ASCENSION, new AscensionBeeEffect());
 
 		apiculture.registerActivityType(ForestryActivityTypes.DIURNAL, new SingleActivityType(0, 12000, ForestryError.NOT_DAY, LightPreference.ANY));